% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaptiveI.R
\name{adaptiveI}
\alias{adaptiveI}
\title{Adaptive method I where only reader sample is resized}
\usage{
adaptiveI(data.interim, samples, randseed = 666, alpha0 = 0.025, beta0 = 0.2)
}
\arguments{
\item{data.interim}{[data.frame] with (nc0+nc1)+nr.1*(nc0+nc1)*2 rows and 4 variables including:

             $readerID: [Factor] w/ nr.1+1 levels "-1", "reader1", "reader2", ...

             $caseID: [Factor] w/ nc0+nc1 levels "negCase1", "negCase2", ..., posCase1", "posCase2", ...

             $modalityID: [Factor] w/ 3 levels "truth", "modility A", "modility B"

             $score: [num] reader score}

\item{samples}{[list] of sample sizes:

              $Nc.0:       [num] fixed number of non-diseased cases

              ($Nc.1):     [num] fixed diseased cases size

              $R.Nc0toNc1: [num] the ratio of the number of non-diseased cases to that of diseased cases (we assume $R.Nc0toNc1>=1)

              $Nr.init:    [num] initial total number of readers

              $Nr.1:       [num] number of readers for the first part of study (before the interim analysis)

              $Nr.max:     [num] maximum number of readers

              ($Nc.0.max:   [num] maximum number of non-diseased cases (to generate the same data as adaptive method II and make comparison); $Nr.expect: [num] expected number of adaptive readers to reach targeted power)}

\item{randseed}{[num] the random seed to indicate the particular doIMRMCfast() intermediate folder/file}

\item{alpha0}{[num] the nominal type I error}

\item{beta0}{[num] 1-beta0 is the targeted power}
}
\value{
result.interim [list] with the updated reader size and critical value, and other intermediate results:

             $nr.adap: [num] updated total reader sample size

             $cv.adap: [num] updated critical value

             $...: other intermediate results
}
\description{
The main function to implement the adaptive method I where only the reader size is adapted after the data for the interim analysis has been collected.
The critical value and the reader sample size are updated simultaneously. The data for interim analysis has been converted to doIMRMC formatted data frame. The cases sizes remain unchanged. Right-sided hypothesis test is considered.
}
