% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{adaptiveII.CondPowerKernel}
\alias{adaptiveII.CondPowerKernel}
\title{Kernel component of calculating the adapted conditional power w.r.t. adaptive method II}
\usage{
adaptiveII.CondPowerKernel(otherparas, nr.adap, nc20.adap)
}
\arguments{
\item{otherparas}{[list] of the following:

                 $nr.1:           [num]  number of readers for the first part of study (before the interim analysis)

                 $nc.10:          [num]  number of non-diseased cases for the first part of the study (before the interim analysis)

                 $nc.11:          [num]  number of diseased cases for the first part of the study (before the interim analysis)

                 $ratio.nc0tonc1: [num]  the ratio of the number of non-diseased cases to that of diseased cases

                 $Vr.1            [num]  estimated variance component w.r.t. the first part of study

                 $Vc.1            [num]  estimated variance component w.r.t. the first part of study

                 $dAUC.moments.1  [list] estimated U-statistics moments

                 $CC1             [num]  constant component for calculating the kernel power component in Formula (xx) of the reference paper

                 $CC2             [num]  constant component for calculating the kernel power component in Formula (xx) of the reference paper}

\item{nr.adap}{[num]  a possible adapted total reader sizes}

\item{nc20.adap}{[num]  a possible adapted non-diseased case sizes}
}
\value{
[list]:

          $CPkernel.adap:  [num] the kernel component of the adapted conditional power

          $cv.adap:        [num] the corresponding adapted critical value

          $...: other potentially necessary outputs
}
\description{
Calculate the kernel component of the adapted conditional power based on:
1. a possible combination of adapted reader and cases sizes for the second part of the study,
2. reader/cases sizes for the first part of the study,
3. the estimates of the population parameters from the first part of the study.
And compute the corresponding adapted critical value.
Note that the adapted diseased case size is proportional to the adapted non-diseased case size (the ratio equals samples$R.Nc0toNc1)
Please refer to xxx paper for more detail.
}
