% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{adaptiveII.condvarR2.varR}
\alias{adaptiveII.condvarR2.varR}
\title{Conditioanl variance and total variance w.r.t. adaptive method II}
\usage{
adaptiveII.condvarR2.varR(nr1, nr2, vr1, vc1, vr2, vc2, vc3)
}
\arguments{
\item{nr1}{[num] number of readers before the interim analysis}

\item{nr2}{[num] number of readers after the interim analysis}

\item{vr1}{[num] Vr based on the numbers of cases before the interim analysis}

\item{vc1}{[num] Vc based on the numbers of cases before the interim analysis}

\item{vr2}{[num] Vr based on the numbers of cases after the interim analysis}

\item{vc2}{[num] Vc based on the numbers of cases after the interim analysis}

\item{vc3}{[num] vc3 equals vc2}
}
\value{
[list] of conditiona variance and the total variance
}
\description{
Calculate the conditional variance of the second part reader-average AUC given the observed first part reader-average AUC
and the total variance where the first and second parts cases sizes are different
for adaptive method II
}
