% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaptiveII.R
\name{adaptiveII}
\alias{adaptiveII}
\title{Adaptive method II where both reader sample and case samples are resized}
\usage{
adaptiveII(data.interim, samples, randseed = 666, alpha0 = 0.025, beta0 = 0.2)
}
\arguments{
\item{data.interim}{[data.frame] with (nc10+nc11)+nr.1*(nc10+nc11)*2 rows and 4 variables including:

             $readerID: [Factor] w/ nr.1+1 levels "-1", "reader1", "reader2", ...

             $caseID: [Factor] w/ nc10+nc11 levels "negCase1", "negCase2", ..., posCase1", "posCase2", ...

             $modalityID: [Factor] w/ 3 levels "truth", "modility A", "modility B"

             $score: [num] reader score}

\item{samples}{[list] of sample sizes:

              $Nc.10:      [num] number of non-diseased cases for the first part of the study (before the interim analysis)

              ($Nc.11):    [num] number of diseased cases size for the first part of the study (before the interim analysis)

              $R.Nc0toNc1: [num] the ratio of the number of non-diseased cases to that of diseased cases (we assume $R.Nc0toNc1>=1)

              $Nc.20:      [num] initial number of non-diseased cases for the second part of the study (after the interim analysis, and we set $Nc.20=$Nc.10)

              ($Nc.21):    [num] initial number of diseased cases for the second part of the study (after the interim analysis, , and we set $Nc.21=$Nc.11)

              $Nc.0.max:   [num] maximum number of non-diseased cases (to generate the same data as adaptive method I and make comparison)

              $Nr.init:    [num] initial total number of readers

              $Nr.1:       [num] number of readers for the first part of study (before the interim analysis)

              $Nr.max:     [num] maximum number of readers

              ($Nr.expect: [num] expected number of adaptive readers to reach targeted power when fixing the cases sizes)}

\item{randseed}{[num] the random seed to indicate the particular doIMRMCfast() intermediate folder/file}

\item{alpha0}{[num] the nominal type I error}

\item{beta0}{[num] 1-beta0 is the targeted power}
}
\value{
result.interim [list] with the updated reader/cases sizes and critical value, and other intermediate results:

             $nr.adap:       [num] updated totoal reader sample size

             $nc.20.adap:    [num] updated non-diseased case sample size for the second part of the study (the updated diseased case sample size is nc.21.adap = $nc.20.adap/$R.Nc0toNc1)

             $cv.adap:       [num] updated critical value

             $...: other intermediate results
}
\description{
The main function to implement the adaptive method II where both the reader and the cases (non-diseased and diseased) sizes are adapted after the data for the interim analysis has been collected.
The critical value and the reader/cases sample sizes are updated simultaneously. The data for interim analysis has been converted to doIMRMC formatted data frame.  Right-sided hypothesis test is considered.
Please refer to xxx paper for more detail.
}
