% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical.var.AUC.R
\name{analytical.var.AUC}
\alias{analytical.var.AUC}
\title{Analytical variance parameters and power in the AUC space give Roe-Metz simulation parameters}
\usage{
analytical.var.AUC(Nr, N0, N1, paras)
}
\arguments{
\item{Nr}{[num] or [vector] (vector of) the number of readers}

\item{N0}{[num] or [vector] (vector of) the number of signal-absent (non-diseased) cases}

\item{N1}{[num] or [vector] (vector of) the number of signal-present (diseased) cases}

\item{paras}{[list] of parameters:

              $mu1: [num] grand average

              $tauA1: [num] modality effect (for the signal-present class)

              $Vr0, $Vtaur0, $Vc0, $Vtauc0, $Vrc0, $Veps0, $b: [num] variance component parameters (see references)}
}
\value{
[list] of 3:

          $var.da:  [num] analytical variance

          $moments: [num] [1 X 8 vector] analytical U-statistics moments

          $pow:     [num] analytical power
}
\description{
Compute the analytical ariance parameters and power in the AUC space give Roe-Metz simulation parameters. Original version was created by Brandon Gallas and Weijie Chen. The original arguments for population parameters are combined as a list argument paras.
The function requires the jiRometz_SPsim2.jar file. Make sure the work directory has this jar file.
}
\references{
Roe and Metz, Acad Radiol 1997, 298-303; Hillis, Acad Radiol 2012, 19:1518.
}
