% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doIMRMCfast.R
\name{doIMRMCfast}
\alias{doIMRMCfast}
\title{A simplified version of doIMRMC() for parallel simulations}
\usage{
doIMRMCfast(dFrame.imrmc, randseed, iMRMCjar = "iMRMC-v4.0.1.jar",
  cleanUp = TRUE)
}
\arguments{
\item{dFrame.imrmc}{[data.frame] the variables contained in this data frame are:

             $readerID: [Factor] w/ nr+1 levels "-1", "reader1", "reader2", ...

             $caseID: [Factor] w/ nc0+nc1 levels "negCase1", "negCase2", ..., posCase1", "posCase2", ...

             $modalityID: [Factor] w/ 3 levels "truth", "modility A", "modility B"

             $score: [num] reader score

 Each row of this data frame corresponds to an observation for every caseID.
 There must be a row corresponding to the truth observation, the readerID for a truth observation is "truth", the modalityID for a truth observation is "truth", the score for a truth observation must be either 0 (non-diseased or signal absent) or 1 (diseased or signal present).}

\item{randseed}{the random seed w.r.t. a particular simulation replicate}

\item{iMRMCjar}{[char] this string identifies the location of the iMRMC.jar file
this jar file can be downloaded from https://github.com/DIDSR/iMRMC/releases
this R program works with version iMRMC-v3p2.jar and later}

\item{cleanUp}{[logi] this logical determines whether or not the iMRMC analysis results}
}
\value{
iMRMCoutput [list] with the following outputs:

           $Ustat: [data.frame] contains the reader-average performance results.
             The analysis results are based on U-statistics.
             Key variables of this data frame are AUCA, AUCB, AUCAminusAUCB and the corresponding
             variances, confidence intervals, degrees of freedom and p-values.

           $moments: [data.frame] contains the 8 U-statistics moments of modality A, modality B.

           $MLEstat: [data.frame] contains the reader-average performance results.
             The analysis results are based on V-statistics, which approximates the true distribution with
             the empirical distribution. The empirical distribution equals the nonparametric MLE
             estimate of the true distribution, which is also equivalent to the ideal bootstrap estimate.
             Key variables of this data frame are AUCA, AUCB, AUCAminusAUCB and the corresponding
             variances, confidence intervals, degrees of freedom and p-values.
}
\description{
Refer to iMRMC package for detailed descriptions of doIMRMC(). This function is a simplefied version of doIMRMC() where: 1. two parameters randseed and hypo ared added to identify different simulation replicates such as to avoid overwriting intermediate folders imrmcDir*** and files input***; 2. only the outputs we need for the adaptive MRMC methods are returned.
}
