% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{fun.vector.c}
\alias{fun.vector.c}
\title{Calculate the vector (c1, c2, c3, c4)}
\usage{
fun.vector.c(nc0, nc1)
}
\arguments{
\item{nc0}{[num] number of non-diseased cases}

\item{nc1}{[num] number of diseasd cases}
}
\value{
[num] vector of (c1, c2, c3, c4)
}
\description{
Calculate the vector (c1, c2, c3, c4) based on the numbers of non-diseased and disease cases
}
