% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryFunctions.R
\name{fun.VrVc}
\alias{fun.VrVc}
\title{Calculate Vr and Vc}
\usage{
fun.VrVc(moments, vector.c)
}
\arguments{
\item{moments}{[num] vector of the eight estimated U-statistics moments}

\item{vector.c}{[num] vector of (c1, c2, c3, c4)}
}
\value{
[list] of Vr and Vc
}
\description{
Calculate Vr and Vc based on the estimated U-statistic moments and the vector (c1, c2, c3, c4)
}
