% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGenerationRM.R
\name{mrmcRMscoresFC}
\alias{mrmcRMscoresFC}
\title{Generate MRMC simulation scores using constrained unequal-variance Roe and Metz model}
\usage{
mrmcRMscoresFC(nr, nc0, nc1, paras)
}
\arguments{
\item{nr}{[num] number of readers}

\item{nc0}{[num] number of signal-absent (non-diseased) cases}

\item{nc1}{[num] number of signal-present (diseased) cases}

\item{paras}{[list] of parameters:

              $mu1: [num] grand average

              $tauA1: [num] modality effect (for the signal-present class)

              $Vr0, $Vtaur0, $Vc0, $Vtauc0, $Vrc0, $Veps0, $b: [num] variance component parameters (see references)}
}
\value{
dFrame.imrmc [data.frame] with (nc0+nc1)+nr*(nc0+nc1)*2 rows and 4 variables including:

             $readerID: [Factor] w/ nr+1 levels "-1", "reader1", "reader2", ...

             $caseID: [Factor] w/ nc0+nc1 levels "negCase1", "negCase2", ..., posCase1", "posCase2", ...

             $modalityID: [Factor] w/ 3 levels "truth", "modility A", "modility B"

             $score: [num] reader score

S10: [num] [nc0 X nr matrix] score w.r.t. modality A, signal absent

S11: [num] [nc1 X nr matrix] score w.r.t. modality A, signal present

S20: [num] [nc0 X nr matrix] score w.r.t. modality B, signal absent

S21: [num] [nc1 X nr matrix] score w.r.t. modality B, signal present
}
\description{
The MRMC simulation scores are generated using the linear mixed effect model (Roe and Metz model) and are fully paired across two modalities (fully-cross design).
}
\references{
Roe and Metz, Acad Radiol 1997, 298-303; Hillis, Acad Radiol 2012, 19:1518.
}
