% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimAdaptiveI.R
\name{sim.adaptiveI}
\alias{sim.adaptiveI}
\title{Implement simulation evaluation of adaptive method I where only reader sample is resized}
\usage{
sim.adaptiveI(samples, paras, randseed = 666, alpha0 = 0.025, beta0 = 0.2)
}
\arguments{
\item{samples}{[list] of sample sizes:

              $Nr.init:    [num] initial total number of readers

              $Nr.1:       [num] number of readers for the first part of study (before the interim analysis)

              $Nr.max:     [num] maximum number of readers

              $Nc.0:       [num] fixed number of non-diseased cases

              $Nc.0.max:   [num] maximum number of non-diseased cases (to generate the same data as adaptive method II and make comparison)

              $R.Nc0toNc1: [num] the ratio of the number of non-diseased cases to that of diseased cases

              ($Nr.expect: [num] expected number of adaptive readers to reach targeted power)}

\item{paras}{[list] of parameters:

              $mu1: [num] grand average

              $tauA1: [num] modality effect (for the signal-present class)

              $b, $Vc0, $Vtauc0, $Vrc0, $Veps0, $Vr0, $Vtaur0: [num] variance component parameters (see references)

              ($varStruc: [factor] variance structure; $AUC_B: [num] base line AUC)}

\item{randseed}{[num] the random seed to generate data}

\item{alpha0}{[num] the nominal type I error}

\item{beta0}{[num] 1-beta0 is the targeted power}
}
\value{
result.interim [list] with the updated reader size, critical value, test decisions and other results:

             $nr.adap:       [num] updated total reader sample size

             $cv.adap:       [num] updated critical value

             $reject.init.z  [num] test decision using wald-test w.r.t. initial sample without interim analysis

             $reject.init.z  [num] test decision using t-test w.r.t. initial sample without interim analysis

             $reject.adap.z  [num] test decision using wald-test w.r.t. adjusted reader size and conventional critical value

             $reject.adap.t  [num] test decision using t-test w.r.t. adjusted reader size and conventional critical value

             $reject.adap.cv [num] test decision using wald-test w.r.t. adjusted reader size and updated critical value

             $...: other intermediate and final results
}
\description{
This is the main function to implement simulations to evaluate the adaptive method I where only the reader size is adapted. The cases sizes remain unchanged. Right-sided hypothesis test is considered
}
\references{
Huang Z, Samuelson F, Tcheuko L, Chen W.Statistical Methods in Medical Research. 2020;29(6):1592-1611.
}
